requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.4
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-timezone-field==7.0
Faker==30.8.2
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
parameterized==0.9.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pytest==8.3.3
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
